requirements.txt
mysqlclient
python-telegram-bot
vk-api
